requirements.txt
mysqlclient

gunicorn

annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.3
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
factory_boy==3.3.3
Faker==37.4.0
idna==3.10
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
python-filter==2.1.0
redis==6.2.0
requests==2.32.4
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
